Finalizado backend

+++ src/main/java/com/apm/jacx/controller/ImagesController.java
@@ -1,15 +1,18 @@
 package com.apm.jacx.controller;
 
+import com.apm.jacx.model.AppUser;
 import com.apm.jacx.model.Image;
+import com.apm.jacx.service.AppUserService;
 import com.apm.jacx.service.ImagesService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.net.URI;
 import java.net.URISyntaxException;
 import java.util.List;
-import java.util.Optional;
 
 @RestController
 @RequestMapping("/api")
@@ -17,33 +20,40 @@ public class ImagesController {
     @Autowired
     private ImagesService imagesService;
 
-    @PostMapping(value = "/image")
-    private ResponseEntity<Image> create(@RequestBody Image image) throws URISyntaxException {
-        Image tmp = imagesService.create(image);
-        return ResponseEntity.created(new URI("api/image/" + tmp.getId())).body(tmp);
-    }
+    @Autowired
+    private AppUserService userService;
 
-    @PutMapping(value = "/image")
-    private ResponseEntity<Image> update(@RequestBody Image image) {
-        Image tmp = imagesService.update(image);
-        return ResponseEntity.ok(tmp);
+    @PostMapping(value = "/image")
+    private ResponseEntity<Image> create(@RequestBody Image image, @RequestHeader(HttpHeaders.AUTHORIZATION) String token) throws URISyntaxException {
+        AppUser appUser = userService.checkToken(token);
+        if (appUser != null) {
+            image.setOwner(appUser);
+            Image tmp = imagesService.create(image);
+            return ResponseEntity.created(new URI("api/image/" + tmp.getId())).body(tmp);
+        } else {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
     }
 
     @GetMapping(value = "/images")
-    private ResponseEntity<List<Image>> getAll() {
-        List<Image> tmp = imagesService.getAll();
-        return ResponseEntity.ok(tmp);
+    private ResponseEntity<List<Image>> getAllByOwner(@RequestHeader(HttpHeaders.AUTHORIZATION) String token) {
+        AppUser appUser = userService.checkToken(token);
+        if (appUser != null) {
+            List<Image> tmp = imagesService.findByOwner(appUser);
+            return ResponseEntity.ok(tmp);
+        } else {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
     }
 
     @DeleteMapping(value = "/image")
-    private ResponseEntity<Void> delete(@RequestBody Image image) {
-        imagesService.delete(image);
-        return ResponseEntity.ok().build();
-    }
-
-    @GetMapping(value = "/image/{id}")
-    private ResponseEntity<Optional<Image>> getByID(@PathVariable("id") Long id) {
-        Optional<Image> tmp = imagesService.findById(id);
-        return ResponseEntity.ok(tmp);
+    private ResponseEntity<Void> delete(@RequestBody Image image, @RequestHeader(HttpHeaders.AUTHORIZATION) String token) {
+        AppUser appUser = userService.checkToken(token);
+        if (appUser != null) {
+            imagesService.delete(image);
+            return ResponseEntity.ok().build();
+        } else {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
     }
 }
